Fix the window resize resulting in animating issue
This patch aims to solve the incorrect animation playing issue when
resizing the window. This results from the animation playing
strategy binded with the animation container re-render and thus lead
to replaying or autoplaying without keypress. To address the issue,
de-couple the animation control from the view and introduce the
presentational component to deal with the transition.